+++ resources/Views/Components/filterInput.php
@@ -1,0 +1,4 @@
+<script src="js/filterDataTable.js"></script>
+<p>
+    <input type="text" id="filterInput" placeholder="<?= __('InsertSearchText') ?>" onkeyup="filterData()" />
+</p>